requirements.txt
# Grok Recruiter - Python Dependencies

# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# HTTP client
httpx>=0.24.0
requests>=2.31.0
requests-oauthlib>=1.3.1  # For X API OAuth 1.0a (DM sending)

# Environment variables
python-dotenv>=1.0.0

# Data visualization
plotly>=5.15.0
matplotlib>=3.7.0

# Dashboard
streamlit>=1.28.0

# Scientific computing (for bandit mocks)
numpy>=1.24.0
scikit-learn>=1.3.0  # For PCA dimensionality reduction

# Logging and utilities
python-json-logger>=2.0.0

# Optional: Neo4j (when Vin's code is ready)
# neo4j>=5.0.0

# Optional: NetworkX (when Vin's code is ready)
# networkx>=3.0

# Optional: SciPy (when Vin's code is ready)
# scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL client

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

